rang buoc CRUD

+++ admin/xulyCRUD_admin/xulyXoa.php
@@ -17,13 +17,24 @@ if(isset($_GET['id_product'])){
 
 if (isset($_GET['id_type'])) {
     $id_type = $_GET['id_type'];
-    $type -> XoaProType($id_type);
+    if ($product->KiemTraIdType($id_type)) {
+        header('location:.././protype.php?messagethongbao=Không thể xóa Type này!!!');
+    }else{
+        $type -> XoaProType($id_type);
     header('location:.././protype.php');
+    }
+    
 }
 if (isset($_GET['id_manu'])) {
    $id_manu = $_GET['id_manu'];
-   $manu->DeleteManu($id_manu);
-   header('location:.././manufacture.php');
+   if ($product->KiemTraIdManu($id_manu)) {
+    header('location:.././manufacture.php?messagethongbao=Không thể xóa Manu này!!!');
+   }
+   else{
+    $manu->DeleteManu($id_manu);
+    header('location:.././manufacture.php');
+   }
+  
 }
 
 ?>
